+++ plugin.php
@@ -22,6 +22,7 @@ require_once NL_DIR.'/migrations/2.0-rc1/Neatline_Migration_20rc1.php';
 require_once NL_DIR.'/migrations/2.0-rc3/Neatline_Migration_20rc3.php';
 require_once NL_DIR.'/migrations/2.0-rc4/Neatline_Migration_20rc4.php';
 require_once NL_DIR.'/migrations/2.0.0/Neatline_Migration_200.php';
+require_once NL_DIR.'/migrations/2.0.2/Neatline_Migration_202.php';
 
 // MODELS
 require_once NL_DIR.'/models/abstract/Neatline_Row_Abstract.php';