+++ tests/phpunit/fixtures/neatline/map/SpatialQueryingEnabledTest.php
@@ -9,8 +9,27 @@
  * @license     http://www.apache.org/licenses/LICENSE-2.0.html
  */
 
-class FixturesTest_NeatlineMapRecordLoading extends Neatline_Case_Fixture
+class FixturesTest_NeatlineMapSpatialQueryingEnabled
+    extends Neatline_Case_Fixture
 {
+
+
+    protected $_isAdminTest = false;
+
+
+    public function testExhibit()
+    {
+
+        $exhibit = $this->_exhibit();
+        $exhibit->spatial_layer = 'OpenStreetMap';
+        $exhibit->spatial_querying = 1;
+        $exhibit->save();
+
+        $this->_writeExhibitMarkupFixture(
+            $exhibit, 'NeatlineMapSpatialQueryingEnabled.html'
+        );
+
+    }
 
 
     public function testRecords()
@@ -26,13 +45,13 @@ class FixturesTest_NeatlineMapRecordLoading extends Neatline_Case_Fixture
         $record2->save();
 
         $this->_writeFixtureFromRoute('neatline/records',
-            'NeatlineMapRecordLoading.two.json'
+            'NeatlineMapSpatialQueryingEnabled.two.json'
         );
 
         $record2->delete();
 
         $this->_writeFixtureFromRoute('neatline/records',
-            'NeatlineMapRecordLoading.one.json'
+            'NeatlineMapSpatialQueryingEnabled.one.json'
         );
 
     }