+++ tests/phpunit/integration/ExhibitsController/admin/EditTest.php
@@ -32,15 +32,16 @@ class ExhibitsControllerTest_AdminEdit extends Neatline_Case_Default
         $imagePath = NL_TEST_DIR.'/mocks/image.jpg';
 
         $exhibit->setArray(array(
-            'title'           => 'Title',
-            'narrative'       => 'Narrative.',
-            'widgets'         => 'Widget1,Widget2',
-            'spatial_layers'  => 'Layer1,Layer3',
-            'spatial_layer'   => 'Layer3',
-            'image_layer'     => $imagePath,
-            'wms_address'     => 'wms.org',
-            'wms_layers'      => 'wms:layer',
-            'public'          => 1
+            'title'             => 'Title',
+            'narrative'         => 'Narrative.',
+            'widgets'           => 'Widget1,Widget2',
+            'spatial_layers'    => 'Layer1,Layer3',
+            'spatial_layer'     => 'Layer3',
+            'image_layer'       => $imagePath,
+            'wms_address'       => 'wms.org',
+            'wms_layers'        => 'wms:layer',
+            'spatial_querying'  => 1,
+            'public'            => 1
         ));
 
         $exhibit->save();
@@ -93,7 +94,11 @@ class ExhibitsControllerTest_AdminEdit extends Neatline_Case_Default
         $this->assertXpath(
             '//input[@name="wms_layers"][@value="wms:layer"]');
 
-        // Public.
+        // Spatial Querying:
+        $this->assertXpath(
+            '//input[@name="spatial_querying"][@checked="checked"]');
+
+        // Public:
         $this->assertXpath(
             '//input[@name="public"][@checked="checked"]');
 
@@ -310,33 +315,35 @@ class ExhibitsControllerTest_AdminEdit extends Neatline_Case_Default
         $imagePath = NL_TEST_DIR.'/mocks/image.jpg';
 
         $this->request->setMethod('POST')->setPost(array(
-            'title'           => 'Title',
-            'slug'            => 'slug',
-            'narrative'       => 'Narrative.',
-            'widgets'         => array('Widget1', 'Widget2'),
-            'spatial_layers'  => array('Layer1', 'Layer2'),
-            'spatial_layer'   => 'Layer1',
-            'image_layer'     => $imagePath,
-            'wms_address'     => 'wms.org',
-            'wms_layers'      => 'wms:layer',
-            'public'          => 1
+            'title'             => 'Title',
+            'slug'              => 'slug',
+            'narrative'         => 'Narrative.',
+            'widgets'           => array('Widget1', 'Widget2'),
+            'spatial_layers'    => array('Layer1', 'Layer2'),
+            'spatial_layer'     => 'Layer1',
+            'image_layer'       => $imagePath,
+            'wms_address'       => 'wms.org',
+            'wms_layers'        => 'wms:layer',
+            'spatial_querying'  => 0,
+            'public'            => 1
         ));
 
         // Submit the form, _reload exhibit.
         $this->dispatch('neatline/edit/'.$exhibit->id);
-        $exhibit = $this->_reload($exhibit);
+        $e = $this->_reload($exhibit);
 
         // Should set fields.
-        $this->assertEquals($exhibit->title,          'Title');
-        $this->assertEquals($exhibit->slug,           'slug');
-        $this->assertEquals($exhibit->narrative,      'Narrative.');
-        $this->assertEquals($exhibit->widgets,        'Widget1,Widget2');
-        $this->assertEquals($exhibit->spatial_layers, 'Layer1,Layer2');
-        $this->assertEquals($exhibit->spatial_layer,  'Layer1');
-        $this->assertEquals($exhibit->image_layer,    $imagePath);
-        $this->assertEquals($exhibit->wms_address,    'wms.org');
-        $this->assertEquals($exhibit->wms_layers,     'wms:layer');
-        $this->assertEquals($exhibit->public,         1);
+        $this->assertEquals($e->title,              'Title');
+        $this->assertEquals($e->slug,               'slug');
+        $this->assertEquals($e->narrative,          'Narrative.');
+        $this->assertEquals($e->widgets,            'Widget1,Widget2');
+        $this->assertEquals($e->spatial_layers,     'Layer1,Layer2');
+        $this->assertEquals($e->spatial_layer,      'Layer1');
+        $this->assertEquals($e->image_layer,        $imagePath);
+        $this->assertEquals($e->wms_address,        'wms.org');
+        $this->assertEquals($e->wms_layers,         'wms:layer');
+        $this->assertEquals($e->spatial_querying,   0);
+        $this->assertEquals($e->public,             1);
 
     }