+++ CHANGELOG.md
@@ -1,4 +1,20 @@
-## v2.0.1 ([commits](https://github.com/scholarslab/Neatline/compare/2.0-rc4...2.0.0)) ~ July 10, 2013
+## v2.0.2 ([commits](https://github.com/scholarslab/Neatline/compare/2.0-rc4...2.0.0)) ~ Pending
+
+#### Added Features
+
+  - Makes it possible to turn spatial querying on and off on a per-exhibit basis. This makes it possible to avoid the added load on the server incurred by real-time querying for exhibits that don't need it (eg, when there are just a handful of records).
+
+#### Changed Features
+
+  - Increases up the size of the "SVG" input textarea to make it easier to see the SVG markup after it's been pasted in.
+
+  - Gets rid of HTML sanitization on exhibit and record models that was scrubbing out `data-neatline-slug` attributes in the markup.
+
+#### Bug Fixes
+
+  - Fixes bug that was making it impossible to paste values from the clipboard into the hex value input boxes for the color fields in the "Style" tab.
+
+## v2.0.1 ([commits](https://github.com/scholarslab/Neatline/compare/2.0.0...2.0.1)) ~ July 10, 2013
 
 #### Bug Fixes